submissions/Team 4/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
torch
faiss-cpu
numpy
tqdm
mwclient
mwparserfromhell
networkx
google-genai
chromadb
requests